+++ src/stories/summary.stories.tsx
@@ -1,9 +1,8 @@
-import React from 'react';
-
 import { Meta, StoryObj } from '@storybook/react';
 
 import { Summary } from '../components/summary/summary';
-import { Color } from '../global/enum';
+
+import { Default } from '../test/summary/default.test';
 
 
 export default {
@@ -20,31 +19,8 @@ export default {
 
 } as Meta<typeof Summary>;
 
-type Story = StoryObj<typeof Summary>;
+export type SummaryStory = StoryObj<typeof Summary>;
 
-const TestDefault = () => {
-
-    return (
-        <>
-            <Summary color={Color.note}>
-                <strong>This is </strong>
-                test.
-                <code>name</code>
-            </Summary>
-            <Summary color={Color.warning}>
-                <strong>This is </strong>
-                test.
-                <code>name</code>
-            </Summary>
-            <Summary color={Color.deep}>
-                <strong>This is </strong>
-                test.
-                <code>name</code>
-            </Summary>
-        </>
-    );
+export { 
+    Default 
 }
-
-export const DefaultTest: Story = {
-    render: () => <TestDefault />
-};